requirements.txt
aiogram==2.11.2
tensorflow==2.4.0
numpy==1.19.5
imageio==2.6.1
matplotlib==3.1.3
Keras==2.4.3
opencv_python==4.5.1.48
keras_contrib==2.0.8
Pillow==8.1.0
